Переписан процессор показа анекдота от указанного пользователя
